@@ -4,7 +4,7 @@
   "description": "React Router Example",
   "main": "dist/application.js",
   "scripts": {
-    "prestart": "webpack && npm run build-css && npm run order-css",
+    "prestart": "webpack --mode=development && npm run build-css && npm run order-css",
     "start": "webpack-dev-server --hot",
     "build-css": "node-sass --include-path scss src/application.scss dist/application.css",
     "order-css": "csscomb src && csscomb dist",
@@ -15,23 +15,24 @@
   "author": "Stephane Moroso",
   "license": "ISC",
   "devDependencies": {
-    "babel-core": "^6.26.0",
-    "babel-eslint": "^8.0.0",
-    "babel-jest": "^21.2.0",
-    "babel-loader": "^7.1.2",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "csscomb": "^4.2.0",
-    "eslint": "^4.6.1",
-    "eslint-plugin-react": "^7.3.0",
-    "jest": "^21.2.1",
-    "node-sass": "^4.5.3",
-    "nodemon": "^1.12.0",
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
-    "react-router-dom": "^4.2.2",
-    "react-test-renderer": "^16.0.0",
-    "webpack": "^3.5.6",
-    "webpack-dev-server": "^2.7.1"
+    "@babel/core": "^7.4.3",
+    "@babel/preset-env": "^7.4.3",
+    "@babel/preset-react": "^7.0.0",
+    "babel-eslint": "^10.0.1",
+    "babel-loader": "^8.0.5",
+    "csscomb": "^4.3.0",
+    "eslint": "^5.16.0",
+    "eslint-plugin-react": "^7.12.4",
+    "jest": "^24.7.1",
+    "node-sass": "^4.11.0",
+    "nodemon": "^1.18.11",
+    "prop-types": "^15.7.2",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
+    "react-router-dom": "^5.0.0",
+    "react-test-renderer": "^16.8.6",
+    "webpack": "^4.30.0",
+    "webpack-cli": "^3.3.1",
+    "webpack-dev-server": "^3.3.1"
   }
 }
